refactor: TSID 및 BaseTimeEntity 구조 통합 및 불필요 코드 제거 (#79)
