+++ src/main/kotlin/kr/wooco/woocobe/place/infrastructure/storage/entity/PlaceJpaEntity.kt
@@ -1,11 +1,11 @@
 package kr.wooco.woocobe.place.infrastructure.storage.entity
 
-import io.hypersistence.utils.hibernate.id.Tsid
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import kr.wooco.woocobe.common.infrastructure.storage.BaseTimeEntity
+import kr.wooco.woocobe.common.infrastructure.storage.Tsid
+import kr.wooco.woocobe.common.infrastructure.storage.entity.BaseTimeEntity
 
 @Entity
 @Table(name = "places")
@@ -26,5 +26,5 @@ class PlaceJpaEntity(
     val name: String,
     @Id @Tsid
     @Column(name = "place_id", nullable = false)
-    val id: Long? = 0L,
+    override val id: Long = 0L,
 ) : BaseTimeEntity()
