+++ src/main/kotlin/kr/wooco/woocobe/place/infrastructure/storage/entity/PlaceOneLineReviewJpaEntity.kt
@@ -1,11 +1,11 @@
 package kr.wooco.woocobe.place.infrastructure.storage.entity
 
-import io.hypersistence.utils.hibernate.id.Tsid
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import kr.wooco.woocobe.common.infrastructure.storage.BaseTimeEntity
+import kr.wooco.woocobe.common.infrastructure.storage.Tsid
+import kr.wooco.woocobe.common.infrastructure.storage.entity.BaseTimeEntity
 
 @Entity
 @Table(name = "place_one_line_reviews")
@@ -18,5 +18,5 @@ class PlaceOneLineReviewJpaEntity(
     val placeId: Long,
     @Id @Tsid
     @Column(name = "place_one_line_review_id")
-    val id: Long? = 0L,
+    override val id: Long = 0L,
 ) : BaseTimeEntity()
