+++ src/main/kotlin/kr/wooco/woocobe/place/infrastructure/storage/entity/PlaceReviewImageJpaEntity.kt
@@ -1,10 +1,11 @@
 package kr.wooco.woocobe.place.infrastructure.storage.entity
 
-import io.hypersistence.utils.hibernate.id.Tsid
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
+import kr.wooco.woocobe.common.infrastructure.storage.Tsid
+import kr.wooco.woocobe.common.infrastructure.storage.entity.BaseTimeEntity
 
 @Entity
 @Table(name = "place_review_images")
@@ -15,5 +16,5 @@ class PlaceReviewImageJpaEntity(
     val placeReviewId: Long,
     @Id @Tsid
     @Column(name = "place_review_images_id")
-    val id: Long? = 0L,
-)
+    override val id: Long = 0L,
+) : BaseTimeEntity()
